Moved Expression Builder to separated class
Parametrized Flow Test and PatternTest

Parametrized Flow Test and PatternTest 2

Successfully parametrized tests
