+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/visitor/pattern/Accumulate.java
@@ -10,6 +10,7 @@ import org.drools.modelcompiler.builder.generator.ModelGenerator;
 import org.drools.modelcompiler.builder.generator.RuleContext;
 import org.drools.modelcompiler.builder.generator.drlxparse.ConstraintParser;
 import org.drools.modelcompiler.builder.generator.drlxparse.DrlxParseResult;
+import org.drools.modelcompiler.builder.generator.expression.ExpressionBuilder;
 import org.drools.modelcompiler.builder.generator.visitor.DSLNode;
 
 class Accumulate implements DSLNode {
@@ -37,7 +38,7 @@ class Accumulate implements DSLNode {
 
             drlxParseResult.accept(success -> {
                 success.setSkipThisAsParam(true);
-                ModelGenerator.processExpression(context, success );
+                ExpressionBuilder.processExpression(context, success );
             });
         }
     }
